requirements.txt
python-telegram-bot==20.8
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.1
pandas==2.1.4
openpyxl==3.1.2